Separating mock and read dataService
